Rename file name dbServer to mockDbServer
